implementing custom hook
